commit sme ecommerce project
