+++ src/main/java/org/mosesidowu/smeecommerce/data/models/ProductCategory.java
@@ -13,4 +13,5 @@ public enum ProductCategory {
     GROCERIES,
     OTHERS;
 
+
 }
