+++ src/main/java/org/mosesidowu/smeecommerce/utils/ProductMapper.java
@@ -3,8 +3,11 @@ package org.mosesidowu.smeecommerce.utils;
 import org.mosesidowu.smeecommerce.data.models.Product;
 import org.mosesidowu.smeecommerce.data.models.ProductCategory;
 import org.mosesidowu.smeecommerce.dtos.requests.CreateProductRequest;
+import org.mosesidowu.smeecommerce.dtos.requests.ProductRequestDTO;
 import org.mosesidowu.smeecommerce.dtos.responses.AllProductsResponse;
 import org.mosesidowu.smeecommerce.dtos.responses.CreateProductResponse;
+import org.mosesidowu.smeecommerce.exception.InvalidCategoryException;
+import org.mosesidowu.smeecommerce.exception.UserException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -51,5 +54,22 @@ public class ProductMapper {
         response.setImageUrl(product.getProductImageUrl());
 
         return response;
+    }
+
+    public static void updateMapperProductResponse(ProductRequestDTO productDTO, Product existingProduct) {
+        existingProduct.setProductName(productDTO.getProductName());
+        existingProduct.setProductDescription(productDTO.getProductDescription());
+        existingProduct.setProductPrice(productDTO.getProductPrice());
+        existingProduct.setProductQuantity(productDTO.getProductQuantity());
+        existingProduct.setProductCategory(safeParseCategory(String.valueOf(productDTO.getProductCategory())));
+    }
+
+
+    public static ProductCategory safeParseCategory(String categoryStr) {
+        try {
+            return ProductCategory.valueOf(categoryStr.trim().toUpperCase());
+        } catch (UserException | NullPointerException e) {
+            throw new InvalidCategoryException("Invalid product category: " + categoryStr);
+        }
     }
 }
